requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes -o requirements.txt
annotated-types==0.7.0
    # via pydantic
asgiref==3.11.1
    # via django
bcrypt==5.0.0
    # via paramiko
boto3==1.42.84
    # via django-ses
botocore==1.42.84
    # via
    #   boto3
    #   s3transfer
certifi==2026.2.25
    # via requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.4.7
    # via
    #   reportlab
    #   requests
cryptography==46.0.6
    # via
    #   paramiko
    #   pyjwt
decorator==5.2.1
    # via fabric
deprecated==1.3.1
    # via fabric
dj-database-url==2.3.0
    # via specbridge
django==5.2.12
    # via
    #   dj-database-url
    #   django-ninja
    #   django-ses
    #   specbridge
django-ninja==1.4.5
    # via specbridge
django-ses==4.7.2
    # via specbridge
fabric==3.2.3
    # via specbridge
gunicorn==23.0.0
    # via specbridge
idna==3.11
    # via requests
invoke==2.2.1
    # via
    #   fabric
    #   paramiko
    #   specbridge
jmespath==1.1.0
    # via
    #   boto3
    #   botocore
lxml==6.0.2
    # via python-docx
packaging==26.0
    # via gunicorn
paramiko==4.0.0
    # via fabric
pillow==12.2.0
    # via reportlab
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.12.5
    # via django-ninja
pydantic-core==2.41.5
    # via pydantic
pyjwt==2.12.1
    # via specbridge
pynacl==1.6.2
    # via paramiko
pypdf==6.9.2
    # via specbridge
python-dateutil==2.9.0.post0
    # via botocore
python-docx==1.2.0
    # via specbridge
python-dotenv==1.2.2
    # via specbridge
reportlab==4.4.10
    # via specbridge
requests==2.33.1
    # via specbridge
s3transfer==0.16.0
    # via boto3
six==1.17.0
    # via python-dateutil
sqlparse==0.5.5
    # via django
typing-extensions==4.15.0
    # via
    #   dj-database-url
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3 ; sys_platform == 'win32'
    # via django
urllib3==2.6.3
    # via
    #   botocore
    #   requests
whitenoise==6.12.0
    # via specbridge
wrapt==2.1.2
    # via deprecated
